Refactor XR plugin
